Creates two functions that can cache the inverse of a matrix
The first function creates a list containing functions to: set a matrix,

get a matrix, set the inverse of the matrix, get the inverse of the
matrix.

The second function shows the inverse of a matrix and if the inverse has
already been calculated, the cache will show the matrix.